requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
openai
httpx
distro
jiter
httpcore